+++ src/theme/prose/badge.ts
@@ -1,0 +1,8 @@
+/**
+ * ProseBadge
+ * ---
+ */
+
+export default {
+  base: 'rounded-full'
+}